Removes the cyclic dependency caused by Lambda
The lambda node needs access to BoardRunner and BoardRunner loads Core.
Core requires Lambda, and boom. Cycle.
